+++ server/src/Documents/Document.ts
@@ -3,8 +3,8 @@ import { STATIC_PREFIX } from "./DocumentsCollection";
 import type { ComplexToken, StructComplexToken } from "../Tokenizer/types";
 import type DocumentsCollection from "./DocumentsCollection";
 
-export type OwnedComplexTokens = { owner: string | null; tokens: ComplexToken[] };
-export type OwnedStructComplexTokens = { owner: string | null; tokens: StructComplexToken[] };
+export type OwnedComplexTokens = { owner?: string; tokens: ComplexToken[] };
+export type OwnedStructComplexTokens = { owner?: string; tokens: StructComplexToken[] };
 
 export default class Document {
   constructor(
@@ -42,13 +42,7 @@ export default class Document {
   }
 
   public getGlobalComplexTokensWithRef(computedChildren: string[] = []): OwnedComplexTokens[] {
-    const localStandardLibDefinitions = this.collection.get("nwscript");
-    return [
-      { owner: this.base ? null : this.uri, tokens: this.complexTokens },
-      ...(localStandardLibDefinitions
-        ? [{ owner: localStandardLibDefinitions.uri, tokens: localStandardLibDefinitions.complexTokens }]
-        : []),
-    ].concat(
+    return [{ owner: this.base ? undefined : this.uri, tokens: this.complexTokens } as OwnedComplexTokens].concat(
       this.children.flatMap((child) => {
         // Cycling children or/and duplicates
         if (computedChildren.includes(child)) {
@@ -92,7 +86,7 @@ export default class Document {
   }
 
   public getGlobalStructComplexTokensWithRef(computedChildren: string[] = []): OwnedStructComplexTokens[] {
-    return [{ owner: this.base ? null : this.uri, tokens: this.structComplexTokens }].concat(
+    return [{ owner: this.base ? undefined : this.uri, tokens: this.structComplexTokens } as OwnedStructComplexTokens].concat(
       this.children.flatMap((child) => {
         // Cycling children or/and duplicates
         if (computedChildren.includes(child)) {