+++ server/src/Providers/Builders/index.ts
@@ -1,5 +1,6 @@
 import CompletionItemBuilder from "./CompletionItemBuilder";
 import HoverContentBuilder from "./HoverContentBuilder";
 import SignatureHelpBuilder from "./SignatureHelpBuilder";
+import SymbolBuilder from "./SymbolBuilder";
 
-export { CompletionItemBuilder, HoverContentBuilder, SignatureHelpBuilder };
+export { CompletionItemBuilder, HoverContentBuilder, SignatureHelpBuilder, SymbolBuilder };